+++ src/components/ScrollerVideo/ScrollerVideo.svelte
@@ -4,7 +4,6 @@
   import Debug from './Debug.svelte';
   import type { Snippet } from 'svelte';
   import { setContext } from 'svelte';
-  import { dev } from '$app/environment';
   import { Tween } from 'svelte/motion';
 
   interface Props {
@@ -202,8 +201,7 @@
       ) {
         scrollerVideo.setVideoPercentage(videoPercentage, {
           jump: false,
-          easing: (t) => t,
-          transitionSpeed: restProps.transitionSpeed,
+          transitionSpeed: restProps.transitionSpeed || 8,
         });
       }
     }
@@ -233,7 +231,7 @@
 <!-- renders Debug component and children foregrounds -->
 {#snippet supportingElements()}
   {#if scrollerVideo}
-    {#if showDebugInfo && dev}
+    {#if showDebugInfo}
       <div class="debug-info">
         <Debug componentState={scrollerVideo.componentState} />
       </div>